frontend/requirements.txt
streamlit==1.40.2
pandas==2.2.3
plotly==5.24.1
azure-storage-blob==12.24.0
azure-cosmos==4.9.0
python-dotenv==1.0.1
azure-identity==1.19.0
requests==2.32.3
numpy==2.1.3
tornado<=6.4.2